--- a/.dockerignore
+++ b/.dockerignore
@@ -9,4 +9,5 @@
 docs-site/src/assets
 node_modules
 !yarn.lock
+!package.json
 !.boltrc.js