--- a/.dockerignore
+++ b/.dockerignore
@@ -11,3 +11,4 @@ node_modules
 !yarn.lock
 !package.json
 !.boltrc.js
+!lerna.json