--- a/Dockerfile
+++ b/Dockerfile
@@ -11,10 +11,6 @@ COPY .boltrc.js .
 COPY yarn.lock .
 RUN rm -rf /app/packages/uikit-workshop
 
-RUN cd packages/twig-renderer && yarn run setup
-RUN cd packages/drupal-twig-extensions && yarn run setup 
-RUN cd packages/core-php && yarn run setup
-
 RUN yarn
 
 CMD node packages/servers/default-server