--- a/Dockerfile
+++ b/Dockerfile
@@ -9,6 +9,7 @@ COPY packages /app/packages
 COPY www /app/www
 COPY .boltrc.js .
 COPY package.json .
+COPY lerna.json .
 COPY yarn.lock .
 RUN rm -rf /app/packages/uikit-workshop