--- a/Dockerfile
+++ b/Dockerfile
@@ -9,10 +9,14 @@ COPY packages /app/packages
 COPY www /app/www
 COPY .boltrc.js .
 COPY package.json .
-COPY lerna.json .
 COPY yarn.lock .
+COPY packages/servers/default-server/package.json .
 RUN rm -rf /app/packages/uikit-workshop
 
 RUN yarn
+
+RUN cd packages/twig-renderer && yarn run setup
+RUN cd packages/drupal-twig-extensions && yarn run setup 
+RUN cd packages/core-php && yarn run setup
 
 CMD node packages/servers/default-server