--- a/Dockerfile
+++ b/Dockerfile
@@ -8,12 +8,11 @@ COPY docs-site /app/docs-site
 COPY packages /app/packages
 COPY www /app/www
 COPY .boltrc.js .
-COPY package.json .
 COPY yarn.lock .
-COPY packages/servers/default-server/package.json .
+COPY lerna.json .
 RUN rm -rf /app/packages/uikit-workshop
 
-RUN yarn
+RUN yarn --cwd packages/servers/default-server --ignore-optional --ignore-platform --ignore-scripts --ignore-engines --skip-integrity-check --production --modules-folder node_modules
 
 RUN cd packages/twig-renderer && yarn run setup
 RUN cd packages/drupal-twig-extensions && yarn run setup