--- a/docs-site/__tests__/server-side-rendering.e2e.js
+++ b/docs-site/__tests__/server-side-rendering.e2e.js
@@ -1,0 +1,17 @@
+module.exports = {
+  'Pattern Lab: Confirm SSR Working': function(browser) {
+    const { testingUrl } = browser.globals;
+    console.log(`global browser url: ${testingUrl}`);
+    browser
+      .url(
+        `${testingUrl}/pattern-lab/patterns/02-components-button-70-server-side-rendering--twig-only/02-components-button-70-server-side-rendering--twig-only.html`,
+      )
+      .waitForElementVisible('bolt-button', 3000)
+      .assert.elementPresent('.c-bolt-button')
+      .saveScreenshot(
+        `screenshots/pattern-lab/web-component-ssr--${browser.currentEnv ||
+          'chrome'}.png`,
+      )
+      .end();
+  },
+};