--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--twig-plus-web-component.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--twig-plus-web-component.twig
@@ -1,4 +1,4 @@
-<p>Demo of server-side rendering a vanilla <code><bolt-button></code> web component with a Twig-generated <code><bolt-icon></code> component inside of it.</p>
+<p>Demo of server-side rendering a vanilla <code>&lt;bolt-button&gt;</code> web component with a Twig-generated <code>&lt;bolt-icon&gt;</code> component inside of it.</p>
 
 {% filter bolt_ssr %}
   <bolt-button>