--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--twig.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--twig.twig
@@ -1,4 +1,4 @@
-<p>Demo of server-side rendering a Twig-generated <code><bolt-button></code> web component.</p>
+<p>Demo of server-side rendering a Twig-generated <code>&lt;bolt-button&gt;</code> web component.</p>
 
 {% filter bolt_ssr %}
    {% include "@bolt-components-button/button.twig" with {