--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component-wo-shadow-dom.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component-wo-shadow-dom.twig
@@ -1,0 +1,9 @@
+{# This demo is also reused for confirming in Nightwatch.js that server-side pre-rendering w/ Twig integration is working as expected #}
+<p>Demo server-side rendering a vanilla <code><bolt-button></code> web component with Shadow DOM manually disabled.</p>
+
+{% filter bolt_ssr %}
+  <bolt-button no-shadow>
+    SSR via pure web component!
+    <bolt-icon name="chevron-right" no-shadow></bolt-icon>
+  </bolt-button>
+{% endfilter %}