--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component-wo-shadow-dom.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component-wo-shadow-dom.twig
@@ -1,5 +1,5 @@
 {# This demo is also reused for confirming in Nightwatch.js that server-side pre-rendering w/ Twig integration is working as expected #}
-<p>Demo server-side rendering a vanilla <code><bolt-button></code> web component with Shadow DOM manually disabled.</p>
+<p>Demo server-side rendering a vanilla <code>&lt;bolt-button&gt;</code> web component with Shadow DOM manually disabled.</p>
 
 {% filter bolt_ssr %}
   <bolt-button no-shadow>