--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component-wo-shadow-dom.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component-wo-shadow-dom.twig
@@ -4,6 +4,6 @@
 {% filter bolt_ssr %}
   <bolt-button no-shadow>
     SSR via pure web component!
-    <bolt-icon name="chevron-right" no-shadow></bolt-icon>
+    <bolt-icon name="chevron-right" slot="after" no-shadow></bolt-icon>
   </bolt-button>
 {% endfilter %}