--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component.twig
@@ -1,0 +1,8 @@
+<p>Demo of server-side rendering a vanilla <code><bolt-button></code> web component. This example will use Shadow DOM when the client-side rendering kicks in (if the browser supports it).</p>
+
+{% filter bolt_ssr %}
+  <bolt-button>
+    SSR via pure web component!
+    <bolt-icon name="chevron-right"></bolt-icon>
+  </bolt-button>
+{% endfilter %}