--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component.twig
@@ -1,4 +1,4 @@
-<p>Demo of server-side rendering a vanilla <code><bolt-button></code> web component. This example will use Shadow DOM when the client-side rendering kicks in (if the browser supports it).</p>
+<p>Demo of server-side rendering a vanilla <code>&lt;bolt-button&gt;</code> web component. This example will use Shadow DOM when the client-side rendering kicks in (if the browser supports it).</p>
 
 {% filter bolt_ssr %}
   <bolt-button>