--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-button-ssr-example--web-component.twig
@@ -3,6 +3,6 @@
 {% filter bolt_ssr %}
   <bolt-button>
     SSR via pure web component!
-    <bolt-icon name="chevron-right"></bolt-icon>
+    <bolt-icon name="chevron-right" slot="after"></bolt-icon>
   </bolt-button>
 {% endfilter %}