--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-server-side-rendering--wc-only.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/button/70-server-side-rendering--wc-only.twig
@@ -1,6 +1,0 @@
-{% filter bolt_ssr %}
-  <bolt-button>
-    SSR via pure web component!
-    <bolt-icon name="chevron-right"></bolt-icon>
-  </bolt-button>
-{% endfilter %}