--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-icon-ssr-example--twig.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-icon-ssr-example--twig.twig
@@ -1,3 +1,5 @@
+<p>Demo of server-side rendering a Twig-generated  <code><bolt-icon></code> web component.</p>
+
 {% filter bolt_ssr %}
   {% include "@bolt-components-icon/icon.twig" with {
     name: "add-open",