--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-icon-ssr-example--web-component.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-icon-ssr-example--web-component.twig
@@ -1,3 +1,5 @@
+<p>Demo of server-side rendering a vanilla <code><bolt-icon></code> web component.</p>
+
 {% filter bolt_ssr %}
   <bolt-icon size="medium" name="customer-service"></bolt-icon>
 {% endfilter %}