--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-icon-ssr-example--web-component.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-icon-ssr-example--web-component.twig
@@ -1,4 +1,4 @@
-<p>Demo of server-side rendering a vanilla <code><bolt-icon></code> web component.</p>
+<p>Demo of server-side rendering a vanilla <code>&lt;bolt-icon&gt;</code> web component.</p>
 
 {% filter bolt_ssr %}
   <bolt-icon size="medium" name="customer-service"></bolt-icon>