--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-server-side-rendering--twig-only.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-server-side-rendering--twig-only.twig
@@ -1,0 +1,7 @@
+{% filter bolt_ssr %}
+  {% include "@bolt-components-icon/icon.twig" with {
+    name: "add-open",
+    background: "circle",
+    size: "medium"
+  } %}
+{% endfilter %}