--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-server-side-rendering--wc-only.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/icon/30-server-side-rendering--wc-only.twig
@@ -1,0 +1,3 @@
+{% filter bolt_ssr %}
+  <bolt-icon size="medium" name="customer-service"></bolt-icon>
+{% endfilter %}