--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/text/30-text-ssr-example.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/text/30-text-ssr-example.twig
@@ -1,3 +1,5 @@
+<p>Demo of server-side rendering a vanilla <bolt-code-snippet><bolt-text></bolt-code-snippet> web component.</p>
+
 {% filter bolt_ssr %}
   <bolt-text headline font-size="xxxlarge">
     Server-side rendered xxxlarge headline (via bolt-text)