--- a/docs-site/src/pages/pattern-lab/_patterns/02-components/text/30-text-ssr-example.twig
+++ b/docs-site/src/pages/pattern-lab/_patterns/02-components/text/30-text-ssr-example.twig
@@ -1,4 +1,4 @@
-<p>Demo of server-side rendering a vanilla <bolt-code-snippet><bolt-text></bolt-code-snippet> web component.</p>
+<p>Demo of server-side rendering a vanilla <code>&lt;bolt-text&gt;</code> web component.</p>
 
 {% filter bolt_ssr %}
   <bolt-text headline font-size="xxxlarge">