--- a/docs-site/src/templates/_navbar-with-version-selector.twig
+++ b/docs-site/src/templates/_navbar-with-version-selector.twig
@@ -1,0 +1,6 @@
+{% extends "@bolt-components-navbar/navbar.twig" %}
+    
+{% block navbar_title %}
+  {{ parent() }}
+  {% include "@bolt-site/version-selector.twig" only %}
+{% endblock %}