--- a/docs-site/src/templates/_page-header.twig
+++ b/docs-site/src/templates/_page-header.twig
@@ -1,5 +1,5 @@
 {% set navbar %}
-  {% embed "@bolt-components-navbar/navbar.twig" with {
+  {% include "@bolt-site/_navbar-with-version-selector.twig" with {
     theme: "xdark",
     width: "auto",
     title: {
@@ -39,21 +39,14 @@
       }
     ],
   } only %}
-    {% block navbar_title %}
-      {{ parent() }}
-      {% include "@bolt-site/version-selector.twig" only %}
-    {% endblock %}
-  {% endembed %}
 {% endset %}
 
-{% embed "@bolt-components-sticky/sticky.twig" with {
+{% include "@bolt-components-sticky/sticky.twig" with {
   attributes: {
     class: [
       "c-bds-layout__header"
     ]
-  }
+  },
+  content: navbar
 } %}
-  {% block sticky_content %}
-    {{ navbar }}
-  {% endblock %}
-{% endembed %}
+