--- a/docs-site/src/templates/_site-head.twig
+++ b/docs-site/src/templates/_site-head.twig
@@ -76,13 +76,17 @@
 
     {# @TODO: wire this up to use Critical CSS #}
     <style>
-      {{ inline( assets["bolt-components-critical-fonts.css"] | default("") ) }}
+      {% if fileExists("@bolt-assets/bolt-components-critical-fonts.css") %}
+        {{ inline( assets["bolt-components-critical-fonts.css"] | default("") ) }}
+      {% endif %}
     </style>
 
     <script>
-      {{ inline( assets["bolt-components-critical-css.js"] | default("") ) }}
-      {{ inline( assets["bolt-components-critical-css-vars.js"] | default("") ) }}
-      {{ inline( assets["bolt-components-critical-fonts.js"] | default("") ) }}
+      {% if fileExists("@bolt-assets/bolt-components-critical-css.js") %}
+        {{ inline( assets["bolt-components-critical-css.js"] | default("") ) }}
+        {{ inline( assets["bolt-components-critical-css-vars.js"] | default("") ) }}
+        {{ inline( assets["bolt-components-critical-fonts.js"] | default("") ) }}
+      {% endif %}
     </script>
 
     {# <script>console.log('Bolt Global Twig variable. In Twig, you can tap into the `bolt` global var to access all sorts of helpful information about the Bolt Design System, such as colors, spacing, and more. Most is exported from Sass; to add more, add this to any Sass file: `@include export-data("myVariables.bolt.json", $my-variables-in-a-sass-map);`, and then you can access it in Twig via `bolt.data.myVariables`. Here is all the data available to the current templates, including the `bolt` var:');</script> #}