--- a/docs-site/src/templates/default.twig
+++ b/docs-site/src/templates/default.twig
@@ -79,30 +79,30 @@
     </div>
 
     <footer role="contentinfo" class="c-bds-layout__footer c-bolt-site__footer">
-      {% embed "@bolt-components-band/band.twig" with {
+      {% set band_content %}
+        {% set gitBranch = 'master' %}
+
+        <nav>
+          <ul class="o-bolt-bare-list" style="flex-flow: row wrap; display: flex; justify-content: space-between; align-items: center;">
+            <li class="u-bolt-margin-right@small">
+              {% include "@bolt-components-headline/text.twig" with {
+                size: "xsmall",
+                weight: "bold",
+                tag: "p",
+                text: "Copyright &copy;" ~ "now"|date("Y") ~ " Pegasystems, Inc."
+              } only %}
+            </li>
+          </ul>
+        </nav>
+      {% endset %}
+
+      {% include "@bolt-components-band/band.twig" with {
         theme: "xlight",
         size: "xsmall",
         fullBleed: false,
         page: page,
+        content: band_content
       } only %}
-        {% block band_content %}
-          {% set gitBranch = 'master' %}
-
-          <nav>
-            <ul class="o-bolt-bare-list" style="flex-flow: row wrap; display: flex; justify-content: space-between; align-items: center;">
-              <li class="u-bolt-margin-right@small">
-                {% include "@bolt-components-headline/text.twig" with {
-                  size: "xsmall",
-                  weight: "bold",
-                  tag: "p",
-                  text: "Copyright &copy;" ~ "now"|date("Y") ~ " Pegasystems, Inc."
-                } only %}
-              </li>
-            </ul>
-          </nav>
-
-        {% endblock %}
-      {% endembed %}
     </footer>
   </div>