--- a/packages/build-tools/create-webpack-config.js
+++ b/packages/build-tools/create-webpack-config.js
@@ -52,9 +52,10 @@ async function createWebpackConfig(buildConfig) {
   let langSuffix = `${config.lang ? '-' + config.lang : ''}`;
 
   let themifyOptions = {
-    watchForChanges: config.watch === true ? true : false,
+    watchForChanges:
+      config.watch === true && config.mode !== 'server' ? true : false,
     classPrefix: 't-bolt-',
-    screwIE11: false,
+    screwIE11: config.mode === 'server' ? true : false,
     fallback: {
       filename: 'bolt-css-vars-fallback',
       jsonDataExport: 'theming-css-vars',
@@ -387,6 +388,11 @@ async function createWebpackConfig(buildConfig) {
       new webpack.DefinePlugin(globalJsData),
       new webpack.NamedModulesPlugin(),
       new CopyWebpackPlugin(config.copy ? config.copy : []),
+    ],
+  };
+
+  if (config.mode !== 'server') {
+    webpackConfig.plugins.push(
       new SassDocPlugin(
         {
           src: `${path.dirname(resolve.sync('@bolt/core'))}/styles/`,
@@ -396,8 +402,8 @@ async function createWebpackConfig(buildConfig) {
           outputPath: config.buildDir,
         },
       ),
-    ],
-  };
+    );
+  }
 
   if (!config.prod && config.webpackDevServer) {
     webpackConfig.plugins.push(