--- a/packages/build-tools/tasks/task-collections.js
+++ b/packages/build-tools/tasks/task-collections.js
@@ -74,7 +74,7 @@ async function compileBasedOnEnvironment() {
   return promiseTasks;
 }
 
-async function clean() {
+async function clean(cleanAll = false) {
   config = config || (await getConfig());
   try {
     let dirs = [];
@@ -119,6 +119,9 @@ async function clean() {
       default:
         dirs = [config.buildDir];
         break;
+    }
+    if (cleanAll) {
+      dirs = [config.wwwDir];
     }
     await internalTasks.clean(dirs);
   } catch (error) {
@@ -167,11 +170,11 @@ async function images() {
   }
 }
 
-async function buildPrep() {
+async function buildPrep(cleanAll = false) {
   config = config || (await getConfig());
   try {
     await getExtraTasks();
-    config.prod ? await clean() : '';
+    config.prod ? await clean(cleanAll) : '';
     await internalTasks.mkDirs();
     await manifest.writeBoltManifest();
     if (