--- a/packages/build-tools/utils/config-store.js
+++ b/packages/build-tools/utils/config-store.js
@@ -33,6 +33,7 @@ async function getDefaultConfig() {
       port: ports[0],
       proxyPort: ports[1],
       ip,
+      mode: configSchema.properties.mode.default,
       env: process.env.NODE_ENV,
       enableCache: configSchema.properties.enableCache.default,
       proxyHeader: configSchema.properties.proxyHeader.default,