## Jira
http://vjira2:8080/browse/BDS-749?filter=-1

## Summary
Adds the capability to server-side pre-render Bolt web components via a new new experimental `@bolt/ssr-server` package. 

## Details
- Updated the `@bolt/components-button` templating logic to conditionally adjust what specific markup gets used. Server-side rendered buttons now output virtually the same markup as the button’s Twig template while client-side rendered buttons retain the same markup as before.
- Updated the `<bolt-text>`, `<bolt-button>`, `<replace-with-children>`, and `<replace-with-grandchildren>` templating logic to better handle client-side hydration when server-side rendered.
- The Bolt build tools have been updated with a new under the hood `mode` config which adjusts some of the the build process‘s behavior (ex. no CSS file’s are compiled to speed up the build times) + provides this contextual info to the the components being compiled and rendered.
    - This means that the client-side JS differs slightly from the server-side JS and is generated as a bolt-global.server.js bundle 
    - This MIGHT ultimately not be necessary however further testing and dev work would be needed to confirm whether or not this is feasible and practical.
- Added an E2E integration test with Nightwatch.js to confirm SSR is working as expected in Pattern Lab
- Added a handful of component examples to test the SSR functionality is working as expected:
    - Button component (ex. 
https://feature-server-side-rendering.boltdesignsystem.com/pattern-lab/patterns/02-components-button--button-server-side-rendering--wo-shadow-dom/02-components-button--button-server-side-rendering--wo-shadow-dom.html, among a few others)
    - Text component 
    - Icon component
- Added a new `cleanAll` optional param to the internal `@bolt/build-tools` clean task -- wired up to be used by Jest to ensure tests are being run in a sanitized environment for improved test consistency
- Refactored and relocated the previous top level `servers` folder to now live in the `packages/servers` folder
- Fixed a small issue encountered with the `@bolt/build-tools` that had been preventing Pattern Lab and the Static Site from getting manually compiled via the CLI (unless specified in the `.boltrc` config)
- Reorganized a few files in the `@bolt/core-php` package to be structured more consistently
- Added the missing `fileExists` and `inlineFile` Twig functions to Bolt Core (which should have been added previously)
- ~~Addressed an issue with the Jest visual regression testing logic which throws an error if a `NOW_TOKEN` env token isn't available locally~~
- Updated the `Dockerfile` to account for the new express server location in the codebase 
- Fixed an unrelated issue with our component Jest test async `describe` calls [throwing Jest warnings](https://travis-ci.com/bolt-design-system/bolt/jobs/184204759)
- Updated several docs site templates as a workaround to fix unrelated upstream Twig updates that had been [breaking the build](https://travis-ci.com/bolt-design-system/bolt/jobs/184204760)

Note:
- While this implementation does include a twig filter, function, and tag integrations to use server-side rendering in Twig, the `{% ssr %}` tag isn't ready for prime time (yet) due to the Twig environment's `context` not being passed along. Perhaps @remydenton has some experience with this...

<details>
  <summary><strong>Usage</strong></summary>
This brand new SSR package currently provides three different interfaces to pre-render web components:

### Option 1. via the CLI (super helpful for testing). 

This Node.js CLI currently accepts two arguments:
    1. The string of HTML to pre-render
    2. Boolean to conditionally enable/disable syntax highlighting (which defaults to `true` when using the CLI directly, `false` when being used for PHP rendering)

```
$ node packages/servers/ssr-server/cli.js '<bolt-button color="secondary">SSR Test</bolt-button>'
```

This returns back the pre-rendered, hydration-friendly markup (with or without syntax highlighting)

![image](https://user-images.githubusercontent.com/1617209/54199749-1a682600-44a0-11e9-8692-178f81173c44.png)

### Option 2. via Twig filter / function
There’s also two ways to use this SSR functionality in Twig via the new `bolt_ssr` Twig filter and `bolt_ssr` Twig function. 

Both of these also use the Node.js CLI interface under the hood with the added bonus of returning back the original pre-rendered HTML as a fallback.
```
{% filter bolt_ssr %}
  <bolt-button color=“secondary”>
    SSR Test
  </bolt-button>
{% endfilter %}
```

### Option 3. via JS using the SSR library itself (experimental)

This SSR functionality can even be used in other JS files via the `renderToString` method on the `@bolt/ssr-server` package.

> Note: currently this method currently requires the developer to manually call the `shutDownSSRServer` method afterwards — hopefully something we can automate in the next iteration.
</details>



<details>
  <summary><strong>Misc Notes</strong></summary>
<br>  
- The first couple of implementations of this were powered by a Google Puppeteer instance however this latest iteration managed to swap out Puppeteer for a lighter weight JSDOM implementation that’s been lightly polyfilled! 🎉
- SSR compile times are averaging around 1 to 1.3 seconds (when no Webpack build is required). Quick enough for a larger POC but not quick enough to globally roll this out without expensive compile times on the server   
</details>


<details>
  <summary><strong>Gotchas / Next Steps</strong></summary>

- **Auto-wire Components Waiting To Load**: we should figure out how we want to create a list of the custom element tags that need to be defined before the SSR logic kicks off: https://github.com/bolt-design-system/bolt/blob/cd72b43574e7d33810c80147edc8cb1613048a38/packages/servers/ssr-server/renderer.js#L42 
- **Cache Busting**: Further work needed to get cache busting added to the SSR build process
- **Async Rendering**: Look into parallelizing / making requests more async
- **Alt Interface for PHP Integtation?** We should look into using web sockets or even using a vanilla server request / response interface.
- **Caching**: making the same request to the same component with the same data on the same version of Bolt should pull straight from memory to speed up build times.
- **Compare SSR Perf**: Further research into performance optimizations -- including doing a benchmark between say, SSR for a Vue.js component compared to our current implementation
- **Test CKEditor Integration**: Test integrating CKEditor example content with this new SSR functionality
- **Combine Webpack Builds?** Look into (potentially) combining the client-side vs server-side webpack builds to eliminate having to compile some of the same components twice
- **Research Further Optimizing Client-side Perf**: Lighter weight JS load in the client = faster SSR response times
- **Combine Multiple Servers**: Out of simplicity, we currently have 3 different Express servers -- the default Express server, the testing server when running Jest tests, and now this new SSR server. At a minimum I could see two of these getting merged together into a single server instance (if not merging together all three)
- **Handle Pre-rendered HTML**: the current implementation doesn't know if the markup provided (or pieces of the markup provided) has been pre-rendered which could lead to unintended rendering quirks and/or extra unnecessary processing 
- **Fragment Cache?** If a button component was being pre-rendered but contained an already pre-rendered icon component, how might we handle that?

</details>


<details>
  <summary><strong>Performance Impact</strong></summary>

**Notes:**
1. This is based on the current implementation which will reuses the `bolt-global.server.js` if a Webpack build had already been done -- this speeds up the initial SSR request.
2. CLI benchmarks were are based on this `bolt-global.server.js` file already existing
3. The current implementation is synchronous -- only one request at a time and PHP waits for the request to complete before continuing.

<hr>

### CLI Performance (per compile):
- Without formatting HTML or syntax highlighting: 1.08s average (or 0.05s faster on average than with Prettier formatting)
- With highlighting + prettier: 1.23s average
- Without syntax highlighting, with HTML formatting: 1.13s average

### Pattern Lab Compile Times (Prod) with SSR enabled (Select Usage)
- Using it in a handful of places (~7 places): ~21.97s total compile time
- Without SSR enabled: ~13.s total compile time

### Global Usage (Tests SSR `<bolt-button>` Globally)
- Pre-rendering all `<bolt-button>` components in the static site generator: ~1 minute, 40s compile time (vs 4s compile time normally)
- Pre-rendering all `<bolt-button>` components in Pattern Lab: 11 minutes, 27.7s minutes (vs 13.2s normally)
</details>

## How to test
- [x] Confirm Travis Build working as expected
- [x] Test running build locally to confirm local compile times of PL are acceptable
- [x] Review the handful of component examples that are wired up to use this new SSR functionality (a few `<bolt-button>` demos, a `<bolt-text>` demo, and a few `<bolt-icon>` demos)