@sghoweri 2019-03-14
> * text/30-server-side-rendering--wc-only.twig
> * icon/30-server-side-rendering--twig-only.twig
> * icon/30-server-side-rendering--wc-only.twig

I've updated all 3 of these component demos to make sure they show up correctly in Pattern Lab's navigation. Taking a look at the two other `<bolt-text>` components to see what's going on