@sghoweri 2019-03-14
> These pages now appear to be blank, regression?:
> 
> * [Text icon variations](https://feature-server-side-rendering.boltdesignsystem.com/pattern-lab/patterns/02-components-text--text-icon-variations/02-components-text--text-icon-variations.html)
> * [Text size and weight variations](https://feature-server-side-rendering.boltdesignsystem.com/pattern-lab/patterns/02-components-text-25-text-size-and-weight-variations/02-components-text-25-text-size-and-weight-variations.html)

I've updated the text w/ icon variations page to hide it completely from the menu until we're ready to bring it back to life (post `<bolt-text>` cleanup most likely).

As for the other text size / weight variations page -- that's been broken for a while due to a few schema case changes from 8 months ago (so totally unrelated). In any case, I've also updated that PL demo to get that back up and running.