@sghoweri 2019-03-14
> Thoughts in general:
> 
> * I would definitely like to see more demos/tests handling pre-rendered/fragments of pre-rendered elements. If there are bugs lurking there, let's get out ahead of them.
Totally agree. The plan is for us to continue to throughly test (plus continue to work on SSR performance) before we move to the next level of testing / iteration.

Also I’ve been thinking that it probably makes the most sense for us to slowly roll out support at the component level once things have been tested / updated accordingly.

For example: button required a couple small updates, text required even fewer, and icon required no updates whatsoever.

> * It seems like performance/speed is going to be a major consideration for making the switch to full SSR. Is SSR in Vue much faster?

It depends. Super tiny simple SSR Vue demos are really quick BUT:
1. They often  require a Node server to be running continuously in the background
2. They have poor support for any libraries or code that expect / require to be running in an actual browser environment (ie. `window is undefined`).

TLDR: I don’t know yet. Performance and reliability are top priorities for this but super steep limitations could also limit the value added.

For what it’s worth, JSDOM is nearly universally the goto “fast” method for handing this so I’m curious to see how fast this current setup could run with a fully optimized JS build. 

> * I'm not crazy about the duplication `isServer` [requires in the JS](https://github.com/bolt-design-system/bolt/pull/1097/files#diff-ad8093a42b9e5fc5b748d20b6e11b078R78). On the other hand, if it means dropping Twig at some point, that's still an improvement. Might be a pattern we can improve upon in the future.

Dev huddle discussion? I’m not a huge fan either BUT this would be our replacement for Twig when we can’t just delete a component’s inner HTML when the JS boots up.

Some components won’t require this but others with more advanced hydration / no js functionality would need something along these lines.

> * Don't totally understand what those SSR PHP files are doing. Looks like some voodoo. Possible dev huddle topic?

Sure - dev huddle chat. 😉