@sghoweri 2019-03-15
> @sghoweri reviewed again and just have a couple more comments:
> 
> * In the [icon example](https://feature-server-side-rendering.boltdesignsystem.com/pattern-lab/patterns/02-components-icon-30-icon-ssr-example--twig/02-components-icon-30-icon-ssr-example--twig.html), it looks like the code snippets aren't rendering properly.
>   ![image](https://user-images.githubusercontent.com/36336/54399260-f40adc00-4693-11e9-9955-042d913a0513.png)
> * In the [text example](https://feature-server-side-rendering.boltdesignsystem.com/pattern-lab/patterns/02-components-text-30-text-ssr-example/02-components-text-30-text-ssr-example.html), there is a JS error
>   ![image](https://user-images.githubusercontent.com/36336/54399329-39c7a480-4694-11e9-8b2a-783f554af9c8.png)
> 
> Neither of these would be show stoppers for approving the PR. At this point just waiting on the `replace-with-children` issue, and even that is a minor issue. If you want to handle it in a follow-up PR, just let me know and I'll approve.

Thanks @danielamorse. Looks like the button component's main JS file got renamed recently so I folded the couple of template logic changes in and removed the old `button.standalone.js` file. I THINK that should solve part of these last couple of things you're flagging here. 

On top of that I fixed two button demos' icons that were missing slots + updated all the SSR demos to make sure all web components being talked about in the demos don't accidentally render 🤦‍♂️  -- that's likely the main issue you're seeing in the first screenshot above (no icon `name` prop defined = no SVG renders).